// src/types/castMemberInterface.ts
// Generated by https://quicktype.io

export interface CastMember {
  adult: boolean;
  also_known_as: string[];
  biography: string;
  birthday: string;
  deathday: null;
  gender: number;
  homepage: null;
  id: number;
  imdb_id: string;
  known_for_department: string;
  name: string;
  place_of_birth: string;
  popularity: number;
  profile_path: string;
}

export interface CombinedCredits {
  cast: CombinedCreditsCast[];
  crew: CombinedCreditsCast[];
  id: number;
}

export interface CombinedCreditsCast {
  adult: boolean;
  backdrop_path: null | string;
  genre_ids: number[];
  id: number;
  original_language: OriginalLanguage;
  original_title?: string;
  overview: string;
  popularity: number;
  poster_path: null | string;
  release_date?: string;
  title?: string;
  video?: boolean;
  vote_average: number;
  vote_count: number;
  character?: string;
  credit_id: string;
  order?: number;
  media_type: MediaType;
  origin_country?: OriginCountry[];
  original_name?: string;
  first_air_date?: string;
  name?: string;
  episode_count?: number;
  department?: Department;
  job?: Job;
}

export enum Department {
  Crew = 'Crew',
  Production = 'Production',
}

export enum Job {
  ExecutiveProducer = 'Executive Producer',
  Presenter = 'Presenter',
  Producer = 'Producer',
  Stunts = 'Stunts',
}

export enum MediaType {
  Movie = 'movie',
  Tv = 'tv',
}

export enum OriginCountry {
  De = 'DE',
  GB = 'GB',
  SE = 'SE',
  Us = 'US',
}

export enum OriginalLanguage {
  En = 'en',
  Fr = 'fr',
  Sv = 'sv',
}